The caster gets its own directory, ahead of its own package
Everything that does not vary by domain moves under lib/caster/: the hooks
interface, the target declaration and its loader, ref resolution and placement,
and the skill document's shape. What stays in commands/cast-mold.ts is this
Foundry's own — the sidecar renderer, the artifact contracts, the verify
manifest, and the nouns a reference row reads in.

A move, not a rewrite. Every function body is byte-identical to the one it
replaces, with two exceptions, both of them the seam asserting itself:

- `resolveMoldRef` loses `moldPath`, a parameter no caller's error message ever
  reached for. It was passed by all four call sites and read by none.
- `loadTargetConfig` takes the target DIRECTORY rather than a repo root and a
  target name, matching every placement function in @galaxy-foundry/cast. Where
  a Foundry keeps its targets is that Foundry's layout, and `casts/` is now
  named once in the command instead of three times in the assembly.

Oracle clean over 47 Molds; 68 cast tests pass; typecheck 0.
